+++ govtool/frontend/src/pages/GovernanceActionDetails.tsx
@@ -134,7 +134,7 @@ export const GovernanceActionDetails = () => {
               <Box data-testid="governance-action-details">
                 <GovernanceActionDetailsCard
                   abstainVotes={
-                    state ? state.abstainVotes : data.proposal.abstainVotes
+                    state ? state.abstainVotes : data.proposal.dRepAbstainVotes
                   }
                   createdDate={
                     state ? state.createdDate : data.proposal.createdDate
@@ -151,7 +151,7 @@ export const GovernanceActionDetails = () => {
                   expiryEpochNo={
                     state ? state.expiryEpochNo : data.proposal.expiryEpochNo
                   }
-                  noVotes={state ? state.noVotes : data.proposal.noVotes}
+                  noVotes={state ? state.noVotes : data.proposal.dRepNoVotes}
                   type={type}
                   label={label}
                   title={title}
@@ -165,8 +165,11 @@ export const GovernanceActionDetails = () => {
                     state ? state?.rationale : data.proposal?.rationale
                   }
                   links={state ? state?.references : data.proposal?.references}
-                  yesVotes={state ? state.yesVotes : data.proposal.yesVotes}
+                  yesVotes={state ? state.yesVotes : data.proposal.dRepYesVotes}
                   govActionId={fullProposalId}
+                  protocolParams={
+                    state ? state.protocolParams : data.proposal.protocolParams
+                  }
                 />
               </Box>
             ) : (