+++ govtool/frontend/src/utils/index.ts
@@ -25,5 +25,6 @@ export * from "./mapDtoToProposal";
 export * from "./numberValidation";
 export * from "./openInNewTab";
 export * from "./removeDuplicatedProposals";
+export * from "./replaceNullValues";
 export * from "./testIdFromLabel";
 export * from "./wait";